Support adapting prepared stmt parameters
This couples the query string (which has parameters) together with how the parameters are actually set on a
java.sql.PreparedStatement. This offers greater flexibility than only being able to configure a custom query string,
and having the parameters and their order fixed.

The existing behavior (setters for create and update queries) is expressed using the new facility, and retaining all
existing behavior.